The History tab partially works, integrated in Swift.

+++ Flickr Viewer/FlickrFetcher.m
@@ -23,9 +23,9 @@
     return [self URLForQuery:@"https://api.flickr.com/services/rest/?method=flickr.places.getTopPlacesList&place_type_id=7"];
 }
 
-+ (NSURL *)URLforPhotosInPlace:(id)flickrPlaceId maxResults:(int)maxResults;
++ (NSURL *)URLforPhotosInPlace:(id)flickrPlaceId maxResults:(NSUInteger)maxResults;
 {
-    return [self URLForQuery:[NSString stringWithFormat:@"https://api.flickr.com/services/rest/?method=flickr.photos.search&place_id=%@&per_page=%d&extras=original_format,tags,description,geo,date_upload,owner_name,place_url", flickrPlaceId, maxResults]];
+    return [self URLForQuery:[NSString stringWithFormat:@"https://api.flickr.com/services/rest/?method=flickr.photos.search&place_id=%@&per_page=%lu&extras=original_format,tags,description,geo,date_upload,owner_name,place_url", flickrPlaceId, (unsigned long)maxResults]];
 }
 
 + (NSURL *)URLforRecentGeoreferencedPhotos;
